packages/laptop-charger/pyproject.toml
[tool.poetry]
name = "laptop-charger"
version = "0.1.0"
description = ""
authors = ["Blake Hamm <blake.j.hamm@gmail.com>"]
readme = "README.md"
packages = [{include = "laptop_charger", from="src"}]

[tool.poetry.dependencies]
python = "^3.11"
python-kasa = "^0.6.2.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-asyncio = "^0.23.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
